+++ CHANGELOG.md
@@ -15,6 +15,7 @@ how this file is formatted and how the process works.
 ### Changed
 - Do not store sync errors for Ciirus properties with bad permissions
 - Use `MCPropertyName` to avoid empty property title during Ciirus sync
+- Do not disable context during Ciirus sync processes
 
 ### Fixed
 - Availabilities worker event name for Ciirus